добавлен Field, в котором основной цикл
